Add world check before resetting detail customization
I added asset thumbnails to this plugin some time ago:
https://github.com/Hazelight/EmmsUI/commit/b565d16375f0b3b3551dd0bbb7cecb9e10ee5297

However, there seems to be an issue with the OnWorldCleanup function. I believe that the asset thumbnails creates a temporary world that gets destroyed after some time. That world destruction causes the OnWorldCleanup to trigger which then clears the customization. With this conditional check I try to prevent that from happening. 

It works for me, but I am not sure if it is safe.